Backend/api_microservices/produit/requirements.txt
Flask==3.0.0
Flask-Cors==4.0.0
snowflake-connector-python==3.5.0
spacy==3.7.2